refactor: remove unnecessary lib dir from rule
